+++ lib/ooxml_parser/common_parser/common_data/paragraph/paragrpah_properties/tabs.rb
@@ -1,3 +1,4 @@
+require_relative 'tabs/paragraph_tab'
 module OoxmlParser
   # Class for working with `w:tabs`
   class Tabs < OOXMLDocumentObject